+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/components/index.ts
@@ -1,1 +1,4 @@
 export * from './case-file-view-folder/case-file-view-folder.component';
+export * from './case-file-view-folder/case-file-view-folder-sort/case-file-view-folder-sort.component';
+
+export * from './shared';