+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-file-view/components/shared/case-file-view-overlay-menu/case-file-view-overlay-menu.component.spec.ts
@@ -1,0 +1,34 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CaseFileViewOverlayMenuComponent } from './case-file-view-overlay-menu.component';
+
+describe('CaseFileViewOverlayMenuComponent', () => {
+  let component: CaseFileViewOverlayMenuComponent;
+  let fixture: ComponentFixture<CaseFileViewOverlayMenuComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ CaseFileViewOverlayMenuComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CaseFileViewOverlayMenuComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+
+  it('should set isOpen to false and emit isOpenChange false when calling closeOverlay', () => {
+    component.isOpen = true;
+    spyOn(component.isOpenChange, 'emit');
+    component.closeOverlay();
+
+    expect(component.isOpen).toBe(false);
+    expect(component.isOpenChange.emit).toHaveBeenCalledWith(false);
+  });
+});