+++ RELEASE-NOTES.md
@@ -1,4 +1,15 @@
 ## RELEASE NOTES
+### Version 5.0.25-case-file-view-error-message-display
+**EUI-6826** Display standard error message if Case File View component is unable to retrieve categories and documents data
+
+### Version 5.0.25-case-file-view-media-viewer-integration
+**EUI-5349** Integrate the Media Viewer component with Case File View
+
+### Version 5.0.25-case-file-view-document-tree-functionalities-v3
+**EUI-7066** Case file view search document exceptions
+**EUI-7160** Case file view hand icon for folder clicking
+**EUI-5458** Case file view document view and folder action
+
 ### Version 5.0.24-case-file-view-folder-document-icons-v3
 **EUI-6974** Case file view folder and document icons display
 
@@ -18,7 +29,7 @@
 ### Version 5.0.22-case-file-view-get-categories-and-documents
 **EUI-5481** Service layer for HTTP GET call to CCD endpoint for retrieving categories and documents
 
-### Version 5.0.24-angular11-upgrade
+### Version 5.0.25-angular11-upgrade
 **EUI-6326** Angular 11 upgrade
 
 ### Version 4.22.0-judicial-user-typeahead